Update ME-Model when user saves it
When the user is redirected to the 'summary' page, there is some initialization of data to fetch the morphologies, traces, etc.
At this point we also execute `set(selectedMEModelIdAtom, meModel['@id']);` to have the id so when the user 'saves' the me-model we can update the created me-model and change a flag of 'validated' to true so it will show up in the table.

<IMG> popup + save me-model button

Closes BBPP134-1698