Fetch resource from ID after select on table
There are differences in the types retrieved by the Elastic search view (used to get a list of resources) and if we fetch the full resource (for some detail page).
Some properties have different shape like brainRegion or some attributes are not indexed.

To simplify it, instead of saving the resource directly as it comes from the list, I save the ID and then I fetch the resource. I know that this involves maybe one or two more calls thru the network but